requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

numpy==1.24.4
scipy==1.10.1
scikit-image==0.21.0
scikit-learn==1.3.2
xgboost==2.1.4

opencv-python-headless==4.10.0.84
imageio==2.33.1
pillow==10.2.0

requests==2.32.3

face-alignment==1.3.5

torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

